Long ids

+++ src/main/scala/petclinic/model.scala
@@ -1,7 +1,7 @@
 package petclinic
 
 sealed abstract class BaseEntity extends Product with Serializable {
-  def id: Int
+  def id: Long
 }
 
 sealed abstract class NamedEntity extends BaseEntity {
@@ -13,11 +13,11 @@ sealed abstract class PersonLike extends NamedEntity {
   def lastName: String
 }
 
-final case class Person(id: Int, name: String, firstName: String, lastName: String)
+final case class Person(id: Long, name: String, firstName: String, lastName: String)
     extends PersonLike
 
 final case class Owner(
-    id: Int,
+    id: Long,
     name: String,
     firstName: String,
     lastName: String,
@@ -26,17 +26,17 @@ final case class Owner(
     telephone: String)
     extends PersonLike
 
-final case class PetType(id: Int, name: String) extends NamedEntity
+final case class PetType(id: Long, name: String) extends NamedEntity
 
 final case class Pet(
-    id: Int,
+    id: Long,
     name: String,
     birthDate: java.util.Date,
-    petTypeId: Int,
-    ownerId: Int)
+    petTypeId: Long,
+    ownerId: Long)
     extends NamedEntity
 
-final case class Speciality(id: Int, name: String) extends NamedEntity
+final case class Speciality(id: Long, name: String) extends NamedEntity
 
-final case class Visit(id: Int, date: java.util.Date, description: String, pet: Pet)
+final case class Visit(id: Long, date: java.util.Date, description: String, pet: Pet)
     extends BaseEntity
