+++ src/test/scala/petclinic/Data.scala
@@ -4,25 +4,25 @@ import scala.collection.SortedMap
 import scala.language.implicitConversions
 
 trait Data {
-  type Owners   = SortedMap[Int, Owner]
-  type PetTypes = SortedMap[Int, PetType]
-  type Pets     = SortedMap[Int, Pet]
+  type Owners   = SortedMap[Long, Owner]
+  type PetTypes = SortedMap[Long, PetType]
+  type Pets     = SortedMap[Long, Pet]
 
   case class DB(owners: Owners, petTypes: PetTypes, pets: Pets)
 
   val petTypes: PetTypes =
     SortedMap(
-      1 -> PetType(1, "cat"),
-      2 -> PetType(2, "dog"),
-      3 -> PetType(3, "lizard"),
-      4 -> PetType(4, "snake"),
-      5 -> PetType(5, "bird"),
-      6 -> PetType(6, "hamster")
+      1L -> PetType(1, "cat"),
+      2L -> PetType(2, "dog"),
+      3L -> PetType(3, "lizard"),
+      4L -> PetType(4, "snake"),
+      5L -> PetType(5, "bird"),
+      6L -> PetType(6, "hamster")
     )
 
   val owners: Owners =
     SortedMap(
-      1 -> Owner(
+      1L -> Owner(
         1,
         "George Franklin",
         "George",
@@ -30,7 +30,7 @@ trait Data {
         "110 W. Liberty St.",
         "Madison",
         "6085551023"),
-      2 -> Owner(
+      2L -> Owner(
         2,
         "Betty Davis",
         "Betty",
@@ -38,7 +38,7 @@ trait Data {
         "638 Cardinal Ave.",
         "Sun Prairie",
         "6085551749"),
-      3 -> Owner(
+      3L -> Owner(
         3,
         "Eduardo Rodriguez",
         "Eduardo",
@@ -46,7 +46,7 @@ trait Data {
         "2693 Commerce St.",
         "McFarland",
         "6085558763"),
-      4 -> Owner(
+      4L -> Owner(
         4,
         "Harold Davis",
         "Harold",
@@ -54,7 +54,7 @@ trait Data {
         "563 Friendly St.",
         "Windsor",
         "6085553198"),
-      5 -> Owner(
+      5L -> Owner(
         5,
         "Peter McTavish",
         "Peter",
@@ -62,9 +62,9 @@ trait Data {
         "2387 S. Fair Way",
         "Madison",
         "6085552765"),
-      6 -> Owner(6, "Jean Coleman", "Jean", "Coleman", "105 N. Lake St.", "Monona", "6085552654"),
-      7 -> Owner(7, "Jeff Black", "Jeff", "Black", "1450 Oak Blvd.", "Monona", "6085555387"),
-      8 -> Owner(
+      6L -> Owner(6, "Jean Coleman", "Jean", "Coleman", "105 N. Lake St.", "Monona", "6085552654"),
+      7L -> Owner(7, "Jeff Black", "Jeff", "Black", "1450 Oak Blvd.", "Monona", "6085555387"),
+      8L -> Owner(
         8,
         "Maria Escobito",
         "Maria",
@@ -72,7 +72,7 @@ trait Data {
         "345 Maple St.",
         "Madison",
         "6085557683"),
-      9 -> Owner(
+      9L -> Owner(
         9,
         "David Schroeder",
         "David",
@@ -80,7 +80,7 @@ trait Data {
         "2749 Blackhawk Trail",
         "Madison",
         "6085559435"),
-      10 -> Owner(
+      10L -> Owner(
         10,
         "Carlos Estaban",
         "Carlos",
@@ -92,19 +92,19 @@ trait Data {
 
   val pets: Pets =
     SortedMap(
-      1  -> Pet(1, "Leo", "2000-09-07", 1, 1),
-      2  -> Pet(2, "Basil", "2002-08-06", 6, 2),
-      3  -> Pet(3, "Rosy", "2001-04-17", 2, 3),
-      4  -> Pet(4, "Jewel", "2000-03-07", 2, 3),
-      5  -> Pet(5, "Iggy", "2000-11-30", 3, 4),
-      6  -> Pet(6, "George", "2000-01-20", 4, 5),
-      7  -> Pet(7, "Samantha", "1995-09-04", 1, 6),
-      8  -> Pet(8, "Max", "1995-09-04", 1, 6),
-      9  -> Pet(9, "Lucky", "1999-08-06", 5, 7),
-      10 -> Pet(10, "Mulligan", "1997-02-24", 2, 8),
-      11 -> Pet(11, "Freddy", "2000-03-09", 5, 9),
-      12 -> Pet(12, "Lucky", "2000-06-24", 2, 10),
-      13 -> Pet(13, "Sly", "2002-06-08", 1, 10)
+      1L  -> Pet(1, "Leo", "2000-09-07", 1, 1),
+      2L  -> Pet(2, "Basil", "2002-08-06", 6, 2),
+      3L  -> Pet(3, "Rosy", "2001-04-17", 2, 3),
+      4L  -> Pet(4, "Jewel", "2000-03-07", 2, 3),
+      5L  -> Pet(5, "Iggy", "2000-11-30", 3, 4),
+      6L  -> Pet(6, "George", "2000-01-20", 4, 5),
+      7L  -> Pet(7, "Samantha", "1995-09-04", 1, 6),
+      8L  -> Pet(8, "Max", "1995-09-04", 1, 6),
+      9L  -> Pet(9, "Lucky", "1999-08-06", 5, 7),
+      10L -> Pet(10, "Mulligan", "1997-02-24", 2, 8),
+      11L -> Pet(11, "Freddy", "2000-03-09", 5, 9),
+      12L -> Pet(12, "Lucky", "2000-06-24", 2, 10),
+      13L -> Pet(13, "Sly", "2002-06-08", 1, 10)
     )
 
   val initialDB = DB(owners, petTypes, pets)
